virtual: BORING: move internal API code into its own package
