+++ pkg/virtual/framework/internalapis/fixtures/namespaces.yaml
