+++ pkg/virtual/framework/internalapis/fixtures/workloadclusters.yaml
